+++ benches/lib.rs
@@ -4,20 +4,9 @@ extern crate cadence;
 
 use test::Bencher;
 
-use cadence::{
-    DEFAULT_PORT,
-    StatsdClient,
-    Counted,
-    Timed,
-    Gauged,
-    Metered,
-    Counter,
-    Timer,
-    Gauge,
-    Meter,
-    NopMetricSink,
-    UdpMetricSink
-};
+use cadence::prelude::*;
+use cadence::{DEFAULT_PORT, StatsdClient, Counter, Timer, Gauge, Meter, NopMetricSink,
+              UdpMetricSink};
 
 
 fn new_nop_client() -> StatsdClient<NopMetricSink> {