+++ src/lib.rs
@@ -71,19 +71,8 @@
 //!
 //! ``` rust,no_run
 //! // Import the client.
-//! //
-//! // Note that we're also importing each of the traits that the client uses to emit
-//! // metircs (Counted, Timed, Gauged, and Metered).
-//! use cadence::{
-//!     Counted,
-//!     Timed,
-//!     Gauged,
-//!     Metered,
-//!     StatsdClient,
-//!     UdpMetricSink,
-//!     DEFAULT_PORT
-//! };
-//!
+//! use cadence::prelude::*;
+//! use cadence::{StatsdClient, UdpMetricSink, DEFAULT_PORT};
 //!
 //! fn main() {
 //!     // Create client that will write to the given host over UDP.
@@ -110,13 +99,12 @@
 //! client instance. This might be useful to you if you'd like to swap out the actual
 //! Cadence client with a dummy version when you are unit testing your code.
 //!
+//! Each of these traits are exported in the prelude module. They are also available
+//! in the main module but aren't typically used like that.
+//!
 //! ``` rust,no_run
-//! use cadence::{
-//!     Counted,
-//!     StatsdClient,
-//!     UdpMetricSink,
-//!     DEFAULT_PORT
-//! };
+//! use cadence::prelude::*;
+//! use cadence::{StatsdClient, UdpMetricSink, DEFAULT_PORT};
 //!
 //!
 //! pub struct User {
@@ -178,15 +166,8 @@
 //!
 //! ``` rust,no_run
 //! use std::io;
-//! use cadence::{
-//!     Counted,
-//!     Timed,
-//!     Gauged,
-//!     Metered,
-//!     StatsdClient,
-//!     MetricSink,
-//!     DEFAULT_PORT
-//! };
+//! use cadence::prelude::*;
+//! use cadence::{StatsdClient, MetricSink, DEFAULT_PORT};
 //!
 //! pub struct MyMetricSink;
 //!