+++ CHANGES.md
@@ -3,12 +3,16 @@
 ## [v0.4.0](https://github.com/tshlabs/cadence/tree/master) - ????-??-??
 * Change name of method for getting metric `&str` representation. The old name
   implied that the instance was consumed which it was not.
+* Create `cadence::prelude` module for easy import of `Counted`, `Timed`,
+  `Gauged`, and `Metered` traits via a glob import. Fixes
+  [#4](https://github.com/tshlabs/cadence/issues/4).
 
 ## [v0.3.0](https://github.com/tshlabs/cadence/tree/0.3.0) - 2016-02-07
 * Change `LoggingMetricSink` log target to `cadence::metrics`.
 * Minor documentation improvements. Fixes [#1](https://github.com/tshlabs/cadence/issues/1).
 * Add benchmarks to test suite.
-* Reduce heap allocations when emitting metrics. Fixes [#3](https://github.com/tshlabs/cadence/issues/3).
+* Reduce heap allocations when emitting metrics. Fixes
+  [#3](https://github.com/tshlabs/cadence/issues/3).
 
 ## [v0.2.1](https://github.com/tshlabs/cadence/tree/0.2.1) - 2015-12-27
 * Change Cadence from MIT license to Apache-2.0 for better compatibility with